refactor: address PRRT_kwDOSJAM6s6fwmpz — split operator-hint parsing into a sibling module
operator_hints.py had grown to 1520 lines, over the 1500-line first-party cap
enforced by tests/unit/test_core_decomposition_maintainability.py. Move the
I/O-free directive parsing helpers (feedback/thread id extraction and the
operator-decision marker bounding) into operator_hint_parsing.py, leaving
operator_hints.py at 1405 lines.
